% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentReadsPE.R
\name{candidate.region}
\alias{candidate.region}
\title{Identify candidate regions}
\usage{
candidate.region(PE.RD, islandDepth, min_cut, max_cut)
}
\arguments{
\item{PE.RD}{A \code{character}. File including sequenced pairs.}

\item{islandDepth}{A \code{numeric}. The minimum depth of candidate regions.}

\item{min_cut}{A \code{numeric}. The minimum region length.}

\item{max_cut}{A \code{numeric}. The maximum region length.}
}
\description{
Identify candidate regions from paired-end sequencing data.
}
