% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeGRanges.R
\name{makeGRanges}
\alias{makeGRanges}
\title{makeGRanges}
\usage{
makeGRanges(obj, type = "fixed", filter = list(delta = c(0, Inf), se = c(0,
  Inf), sigmaSqF = c(0, Inf), sigmaSqR = c(0, Inf), score = c(0, Inf)),
  length = 100)
}
\arguments{
\item{obj}{A \code{picsList} object, as returned by \code{PICS}.}

\item{type}{A \code{character}. The type of intervals to be created. The
valid types are  "bed', "wig" and "fixed". See details section for more.}

\item{filter}{A \code{list} of filters to be used before computing the FDR.
By default, all regions are included.}

\item{length}{A \code{numeric}. The length of the intervals when using
 "fixed" type.}
}
\value{
A \code{GRanges} object
}
\description{
Generate a \code{GRanges} objects from a \code{picsList} object.
}
\details{
"bed" will generate intervals from the forward peak max to the reverse peak
max. "wig" will generate a density profile for the forward and reverse reads.
"bed" and "wig" types should be used to be exported to wig/bed files that can
be used with the UCSC genome browser. "fixed" corresponds to the binding site
estimates +/-3*length. "bed" and "wig" files can be exported using the
\code{export} function of the \code{rtracklayer} package.
}
\author{
Renan Sauteraud
}
\seealso{
\code{\link{picsFDR}}
}

