% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesOutput.R
\name{makeGRangesOutput}
\alias{makeGRangesOutput}
\title{Export a PICS object to GRanges}
\usage{
makeGRangesOutput(
  obj,
  type = "fixed",
  length = 100,
  filter = list(delta = c(0, Inf), se = c(0, Inf), sigmaSqF = c(0, Inf), sigmaSqR = c(0,
    Inf), score = c(0, Inf))
)
}
\arguments{
\item{obj}{A \code{PICS} object. The output of the \code{PICS} function.}

\item{type}{A \code{character}. One of "fixed", "bed", "ci" or "wig".}

\item{length}{A \code{numeric}. The length of the region around the center 
(mu). Only used when type = "fixed.}

\item{filter}{A \code{list}. Additional filtering options.}
}
\value{
A \code{GRanges} object.
}
\description{
Export a PICS object to GRanges
}
\seealso{
PICS
}
