% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{pics-class}
\alias{as.data.frame.picsList}
\alias{chromosome}
\alias{chromosome,pics-method}
\alias{delta}
\alias{delta,pics-method}
\alias{mu}
\alias{mu,pics-method}
\alias{pics-class}
\alias{score,pics-method}
\alias{scoreForward}
\alias{scoreForward,pics-method}
\alias{scoreReverse}
\alias{scoreReverse,pics-method}
\alias{se}
\alias{se,pics-method}
\alias{show,pics-method}
\alias{sigmaSqF}
\alias{sigmaSqF,pics-method}
\alias{sigmaSqR}
\alias{sigmaSqR,pics-method}
\alias{w}
\alias{w,pics-method}
\title{pics class}
\description{
This class store the result of a \code{PICS} analysis: The estimated location
of a binding event, its score and the chromosome it belongs to.
}
\section{Slots}{

\describe{
\item{\code{estimates}}{A \code{list} of \code{numeric}. Coordinates of the binding
 event, variance and standard error. Contains the following elements:
\describe{
  \item{w}{The mixture weights. The proportion of reads of the region
  contributing to each binding event.}
  \item{mu}{The binding site positions.}
  \item{delta}{The average DNA fragment lengths. It corresponds to the
   distance between the maxima of the forward and reverse read position
   densities.}
  \item{sigmaSqF}{The variance parameters for the forward distribution.}
  \item{sigmaSqR}{The variance parameters for the reverse distribution.}
  \item{seMu}{The standard errors for mu.}
  \item{seMuF}{The standard errors for muF.}
  \item{seMuR}{The standard errors for muR.}
}}

\item{\code{score}}{A \code{numeric}. The score for the binding event.}

\item{\code{scoreF}}{A \code{numeric}. The score of the forward reads for the binding
event.}

\item{\code{scoreR}}{A \code{numeric}. The score of the reverse reads for the binding
event.}

\item{\code{Nmerged}}{A \code{numeric}. The number of peaks that got merged (in case
of boundary issue).}

\item{\code{converge}}{A \code{logical}. TRUE if the EM has converged.}

\item{\code{range}}{A \code{numeric}. The range of the binding event.}

\item{\code{chr}}{A \code{character}. The chromosome where the binding event is located.}
}}
\author{
Xuekui Zhang, Arnaud Droit, Raphael Gottardo, Renan Sauteraud
}
\seealso{
\code{\linkS4class{picsList}}, \code{\link{PICS}}
}

