% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picsClass.R
\name{pics-generics}
\alias{pics-generics}
\alias{minRange}
\alias{maxRange}
\alias{score}
\alias{scoreReverse}
\alias{scoreForward}
\alias{chromosome}
\alias{se}
\alias{seF}
\alias{seR}
\alias{sigmaSqF}
\alias{sigmaSqR}
\alias{delta}
\alias{mu}
\alias{w}
\alias{K}
\alias{code}
\alias{score,data.frame-method}
\alias{scoreReverse,data.frame-method}
\alias{scoreForward,data.frame-method}
\alias{chromosome,data.frame-method}
\alias{se,data.frame-method}
\alias{seF,data.frame-method}
\alias{seR,data.frame-method}
\alias{sigmaSqF,data.frame-method}
\alias{sigmaSqR,data.frame-method}
\alias{delta,data.frame-method}
\alias{mu,data.frame-method}
\title{Generics associated with pics classes}
\usage{
minRange(x, ...)

maxRange(x, ...)

score(x, ...)

scoreReverse(x, ...)

scoreForward(x, ...)

chromosome(x, ...)

se(x, ...)

seF(x, ...)

seR(x, ...)

sigmaSqF(x, ...)

sigmaSqR(x, ...)

delta(x, ...)

mu(x, ...)

w(x, ...)

K(x, ...)

code(x, ...)

\S4method{score}{data.frame}(x)

\S4method{scoreReverse}{data.frame}(x)

\S4method{scoreForward}{data.frame}(x)

\S4method{chromosome}{data.frame}(x)

\S4method{se}{data.frame}(x)

\S4method{seF}{data.frame}(x)

\S4method{seR}{data.frame}(x)

\S4method{sigmaSqF}{data.frame}(x)

\S4method{sigmaSqR}{data.frame}(x)

\S4method{delta}{data.frame}(x)

\S4method{mu}{data.frame}(x)
}
\arguments{
\item{x, ...}{Object and argumemts passed to the methods.}
}
\description{
These generics are used in methods of \code{pics}, \code{picsError} and
\code{picsList} classes. See class help pages for method documentation.
}
\section{Functions}{
\itemize{
\item \code{minRange}: Start accessor

\item \code{maxRange}: End accessor

\item \code{score}: Score accessor

\item \code{scoreReverse}: Reverse score accessor

\item \code{scoreForward}: Forward score accesor

\item \code{chromosome}: Chromosome accessor

\item \code{se}: se accessor

\item \code{seF}: Forward se accessor

\item \code{seR}: Reverse se accessor

\item \code{sigmaSqF}: sigmaSqF accessor

\item \code{sigmaSqR}: sigmaSqR accessor

\item \code{delta}: delta accessor

\item \code{mu}: mu accessor

\item \code{w}: w accessor

\item \code{K}: K accessor

\item \code{code}: Return error codes

\item \code{score,data.frame-method}: Score accessor

\item \code{scoreReverse,data.frame-method}: Reverse score accessor

\item \code{scoreForward,data.frame-method}: Forward score accessorse

\item \code{chromosome,data.frame-method}: chromosome accessor

\item \code{se,data.frame-method}: se accessor

\item \code{seF,data.frame-method}: seF accessor

\item \code{seR,data.frame-method}: seR accessor

\item \code{sigmaSqF,data.frame-method}: sigmaSqF accessor

\item \code{sigmaSqR,data.frame-method}: sigmaSqR accessor

\item \code{delta,data.frame-method}: delta accessor

\item \code{mu,data.frame-method}: mu accessor
}}

