% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/PICS-package.R
\docType{package}
\name{PICS-package}
\alias{PICS-package}
\title{PICS: Probabilistic Inference for ChIP-Seq}
\description{
This package estimates transcription factors binding sites position
}

