% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/PICS.R
\name{PICS}
\alias{PICS}
\title{Estimation of binding site position}
\usage{
PICS(segReadsList, paraEM = NULL, paraPrior = NULL, nCores = 1)
}
\arguments{
\item{segReadsList}{A \code{segreadsList} object. Contains the segmentation
of the genome in candidate regions}

\item{paraEM}{A \code{list}. List of parameters for the EM algorithm.
\code{setParaEM} can be used to modify specific parameters.}

\item{paraPrior}{A \code{list}. Prior for the data. \code{setParaPrior} can
be used to modify specific parameters.}

\item{nCores}{An \code{integer}. The number of cores that should be used in
 parallel by the function.}
}
\value{
A \code{picsList}
}
\description{
Estimates the position of binding sites in the given segReadsList object.
}
\author{
Renan Sauteraud
}
\seealso{
\code{\link{setParaEM}}, \code{\link{setParaPrior}}
}

