% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{picsError-class}
\alias{chromosome,picsError-method}
\alias{delta,picsError-method}
\alias{mu,picsError-method}
\alias{picsError-class}
\alias{score,picsError-method}
\alias{scoreForward,picsError-method}
\alias{scoreReverse,picsError-method}
\alias{se,picsError-method}
\alias{show,picsError-method}
\alias{sigmaSqF,picsError-method}
\alias{sigmaSqR,picsError-method}
\alias{w,picsError-method}
\title{picsError class}
\description{
This object is used to return an error code when the PICS function failed to
return a valid set of estimates for a candidate regions. This could be due to
non-convergence of the EM algorithm, a singular information matrix, or a
number of reads below the limit specified by the user. All of these are
typically due to too few reads in the region and do not affect the rest of
the analysis, as such regions would most likely be labelled as false
positives.
}
\author{
Xuekui Zhang, Arnaud Droit, Raphael Gottardo, Renan Sauteraud
}
\seealso{
\code{\linkS4class{picsList}}, \code{\link{PICS}}
}

