% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/picsFDR.R
\name{picsFDR}
\alias{picsFDR}
\title{Estimate the FDR.}
\usage{
picsFDR(picsIP, picsCont, filter = list(delta = c(0, Inf), se = c(0, Inf),
  sigmaSqF = c(0, Inf), sigmaSqR = c(0, Inf)))
}
\arguments{
\item{picsIP}{An object of class \code{pics} or \code{picsList} containing
the informations for the IP reads}

\item{picsCont}{An object of class \code{pics} or \code{picsList} containing
the informations for the control reads}

\item{filter}{filter}
}
\value{
A 3 columns \code{data.frame} with the following columns: FDR,
score, N.
}
\description{
Estimate the false detection rate for an object of class \code{pics} or
\code{picsList}.
}
\author{
Xuekui Zhang
}
\seealso{
\code{\linkS4class{picsList}} \code{\linkS4class{pics}}
}
\keyword{functions}

