% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{picsList-class}
\alias{[,picsList,ANY,ANY-method}
\alias{[[,picsList,ANY,ANY-method}
\alias{chromosome,picsList-method}
\alias{delta,picsList-method}
\alias{length,picsList-method}
\alias{mu,picsList-method}
\alias{picsList}
\alias{picsList-class}
\alias{plot,picsList,picsList-method}
\alias{score,picsList-method}
\alias{scoreForward,picsList-method}
\alias{scoreReverse,picsList-method}
\alias{se,picsList-method}
\alias{seF,picsList-method}
\alias{seR,picsList-method}
\alias{show,picsList-method}
\alias{sigmaSqF,picsList-method}
\alias{sigmaSqR,picsList-method}
\alias{w,picsList-method}
\title{picsList class}
\description{
This class stores the result of a \code{PICS} analysis. It consists in a list
of all binding events identified held by \code{pics} objects and some
additional information regarding the parameters used to generate this object.
}
\section{Slots}{

\describe{
\item{\code{List}}{A \code{list} of \code{pics} objects. The list of predicted
binding events.}

\item{\code{paraPrior}}{A \code{list} of \code{numeric}. The list of prior parameters.}

\item{\code{paraEM}}{A \code{list}. The list of parameters used in the EM algorithm
of the \code{PICS} analysis.}

\item{\code{minReads}}{A \code{list} of \code{integer}. The minumum number of reads
per peak and per region.}

\item{\code{N}}{A \code{numeric}. The number of reads used in the analysis.}

\item{\code{Nc}}{A \code{numeric}. The number of control reads used in the analysis.}
}}
\section{Methods}{

\describe{
 \item{show}{signature(object = "segReadsList"): Print details about the
  object.}
 \item{length}{signature(x = "segReadsList"): The length of the \code{List}
  in the object. The number of candidate regions identified.}
}
}
\author{
Xuekui Zhang, Arnaud Droit, Raphael Gottardo, Renan Sauteraud
}
\seealso{
\code{\linkS4class{pics}}, \code{\linkS4class{picsError}}, \code{\link{PICS}}
}

