% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethods.R
\name{plot-pics}
\alias{plot-pics}
\alias{plot,pics,segReads-method}
\alias{plot,picsError,segReads-method}
\alias{plot,picsList,segReadsList-method}
\title{Plot methods for PICS objects}
\usage{
\S4method{plot}{pics,segReads}(
  x,
  y,
  addKernel = FALSE,
  addNucleosome = FALSE,
  addSe = TRUE,
  main = NULL,
  ...
)

\S4method{plot}{picsError,segReads}(x, y, addKernel = FALSE, main = NULL, ...)

\S4method{plot}{picsList,segReadsList}(
  x,
  y,
  regionIndex = NULL,
  addKernel = FALSE,
  addNucleosome = FALSE,
  addSe = TRUE,
  main = NULL,
  ...
)
}
\arguments{
\item{x, y}{Objects}

\item{addKernel}{Add kernel density estimate to the plot.}

\item{addNucleosome}{Add a nucleosome track to the plot.}

\item{addSe}{Add standard error to the plot.}

\item{main}{Main title.}

\item{...}{Arguments to be passed to the plot method.}

\item{regionIndex}{Add region index to the plot.}
}
\description{
Methods to plot \code{pics} and \code{segReads} objects and derived classes.
}
\section{Functions}{
\itemize{
\item \code{plot,pics,segReads-method}: Plot method for \code{pics} and \code{segReads}

\item \code{plot,picsError,segReads-method}: Plot method for \code{picsError} and \code{segReads}

\item \code{plot,picsList,segReadsList-method}: Plot method for \code{picsList} and \code{segReadsList}
}}

