% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segmentPICS.R
\name{segmentPICS}
\alias{segReadsGeneric}
\alias{segmentPICS}
\title{Segment the genome into candidate regions}
\usage{
segmentPICS(data, dataC = NULL, map = NULL, minReads = 2,
  minReadsInRegion = 3, jitter = FALSE, maxLregion = 0,
  minLregion = 100)
}
\arguments{
\item{data}{A file or object containing the IP reads.
See details for more information on how to set up the data.}

\item{dataC}{A file or object containing the control
reads. Set to NULL by default, i.e. no control.}

\item{map}{A file or object containing the mappability profiles. Set to
NULL by default, i.e. no profiles.}

\item{minReads}{The minimum number of F/R reads to be present in the sliding
window.}

\item{minReadsInRegion}{The minimum number of F/R reads to be present in the
region.}

\item{jitter}{A logical value stating whether some noise should be added to
the read locations. This is recommended if the read positions have lots of
duplicates.}

\item{maxLregion}{The maximum length.}

\item{minLregion}{The minimum length.}
}
\value{
An object of class \code{segReadsList} containing the results for
all regions pre-processed.
}
\description{
Pre-process bidirectional aligned reads data from a single ChIP-Seq
experiment to detect candidate regions with a minimum number of forward and
reverse reads. These candidate regions will then be processed by PICS.
}
\details{
The input of data, dataC and map can be an object of class \code{GRanges} or
\code{GAlignments}. Or a bed or bam file.
}
\examples{
# Read data
datadir <- system.file("extdata", package="PICS")
dataIP <- file.path(datadir, "Treatment_tags_chr21_sort.bed")
dataCont <- file.path(datadir, "Input_tags_chr21_sort.bed")
map <- file.path(datadir, "mapProfileShort.bed")

seg<-segmentPICS(dataIP, dataC = dataCont, map = map, minReads = 1)
}
\author{
Renan Sauteraud
}
\references{
X. Zhang, G. Robertson, M. Krzywinski, K. Ning, A. Droit, S.
Jones, and R. Gottardo, PICS: Probabilistic Inference for ChIP-seq
arXiv, 0903.3206, 2009.
}
\seealso{
\code{\linkS4class{segReadsList}}
}
\keyword{functions}

