% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{segReads-class}
\alias{segReads}
\alias{segReads-class}
\alias{show,segReads-method}
\title{segReads class}
\description{
This class stores a candidate region, as determined by a segmentation function.
It contains information regarding the reads that contributes to this region.
Usually, \code{segReads} are part of a \code{segReadsList} object, to be used
as input for the \code{PICS} and \code{PING} function.
}
\section{Slots}{

\describe{
\item{\code{yF}}{A \code{numeric}. The start position of forward reads in the
candidate region.}

\item{\code{yR}}{A \code{numeric}. The start position of reverse reads in the
candidate region.}

\item{\code{cF}}{A \code{numeric}. The start position of forward reads in the control.}

\item{\code{cR}}{A \code{numeric}. The start position of reverse reads in the control.}

\item{\code{map}}{A \code{matrix}. Mappability profile}

\item{\code{chr}}{A \code{character}. The chromosome on which the candidate region is
 located.}
}}
\author{
Xuekui Zhang, Arnaud Droit, Raphael Gottardo, Renan Sauteraud
}
\seealso{
\code{\linkS4class{segReadsList}}, \code{\link{segmentPICS}}, \code{\link{PICS}}
}

