% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segReadsGeneric.R
\name{segReadsGeneric}
\alias{segReadsGeneric}
\title{Perform genome segmentation depending}
\usage{
segReadsGeneric(
  data,
  dataC = NULL,
  map = NULL,
  minReads = 2,
  minReadsInRegion = 3,
  jitter = FALSE,
  maxLregion = 0,
  minLregion = 100,
  step = 20,
  width = 250,
  package = "PICS"
)
}
\arguments{
\item{data}{A \code{GRanges} object containing the IP reads. See details for
more information on how to set up the data.}

\item{dataC}{A \code{GRanges} object containing the control reads. Set to NULL
by default, i.e. no control.}

\item{map}{A \code{GRanges} object containing the mappability profiles. Set
to NULL by default, i.e. no profiles.}

\item{minReads}{A \code{numeric}. The minimum number of F/R reads to be 
present in the sliding window.}

\item{minReadsInRegion}{A \code{numeric}. The minimum number of F/R reads
to be present in the region.}

\item{jitter}{A \code{logical} value stating whether some noise should be 
added to the read locations. This is recommended if the read positions have
 lots of duplicates.}

\item{maxLregion}{A \code{numeric}. The maximum length.}

\item{minLregion}{A \code{numeric}. The minimum length.}

\item{step}{A \code{numeric}. The increment of the sliding window.}

\item{width}{A \code{numeric}. The width of the region.}

\item{package}{A \code{character}. "PICS" or "PING"}
}
\description{
Perform genome segmentation depending
}
