% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{segReadsList-class}
\alias{[,segReadsList,ANY,ANY-method}
\alias{[[,segReadsList,ANY,ANY-method}
\alias{length,segReadsList-method}
\alias{segReadsList}
\alias{segReadsList-class}
\alias{show,segReadsList-method}
\title{segReadsList class}
\description{
This class stores the results of \code{segmentPICS}, the list of all candidate
regions to be used as input in \code{PICS} or \code{PING}. On top of
holding all the \code{segReads}, it contains additional information regarding
the parameters used to generate this object.
}
\section{Slots}{

\describe{
\item{\code{List}}{A \code{list} of \code{segReads} objects. The list of candidate
regions.}

\item{\code{paraSW}}{A \code{list} of \code{numeric}. A list of parameters used in
the segmentation function.}

\item{\code{N}}{A \code{numeric}. The number of reads used as input of the
segmentation function.}

\item{\code{Nc}}{A \code{numeric}. The number of control reads used as input of the
 segmentation function.}
}}
\note{
The segmentation functions are \code{segmentPICS} and \code{segmentPING}
}
\section{Methods}{

\describe{
 \item{show}{signature(object = "segReadsList"): Print details about the
  object.}
 \item{length}{signature(x = "segReadsList"): The length of the \code{List}
  in the object. The number of candidate regions identified.}
}
}
\author{
Xuekui Zhang, Arnaud Droit, Raphael Gottardo, Renan Sauteraud
}
\seealso{
\code{\linkS4class{segReads}}, \code{\link{segmentPICS}}, \code{\link{PICS}}
}

