\name{segReadsListPE}
\alias{segReadsListPE}
\alias{segReadsListPE-class}

\title{List of segReadsPE objects}

\section{Extends}{
Class \code{\linkS4class{segReadsList}}, directly.
}

\author{Xuekui Zhang}

\seealso{
  \code{\linkS4class{segReadsPE}}
  \code{\linkS4class{segReadsList}}
}

\keyword{classes}