\name{segReadsPE}
\alias{segReadsPE}
\alias{segReadsPE-class}

\title{Class to store post-segmentation result}

\description{
This class stores the information of the segmentation performed by \code{segmentPING}. It is used as the input of the \code{PING} function.
}

\section{Extends}{
Class \code{\linkS4class{segReadsList}}, directly.
}

\author{Xuekui Zhang}

\seealso{
  \code{\linkS4class{segReads}}
  \code{\linkS4class{segReadsListPE}}
}

\keyword{classes}