% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPath.R
\docType{methods}
\name{selectPath,FitGene-method}
\alias{selectPath}
\alias{selectPath,FitGene-method}
\title{pathway selection using LASSO}
\usage{
\S4method{selectPath}{FitGene}(object, seed = 123)
}
\arguments{
\item{object}{results of selectGene step}

\item{seed}{random seed that was set,  default = 123.}
}
\description{
LASSO step for pathway selection. To identify patient subgroups based on pathway level risk profile, we fit a LASSO-Penalized Cox regression on scores derived from all the pathways.
}
\examples{
data(TCGA)
prefilter.results=prefilter( data=TCGA$geneexpr, time=TCGA$t, status=TCGA$d, plist=TCGA$pathList )
gene.results=selectGene( prefilter.results )
path.results=selectPath( gene.results )
}

