\name{show}
\Rdversion{1.1}
\docType{methods}
\alias{show-methods}
\alias{show}
\alias{show,pics-method}
\alias{show,picsError-method}
\alias{show,picsList-method}
\alias{show,segReads-method}
\alias{show,segReadsList-method}
\title{show PICS } 

\description{
This methods show the objects of PICS
}

\usage{
\S4method{show}{pics}(object)
\S4method{show}{picsError}(object)
\S4method{show}{picsList}(object)
\S4method{show}{segReads}(object)
\S4method{show}{segReadsList}(object)
}

\arguments{
\item{object}{Object returned from \code{\link{pics}} .}
}

\details{
List of the slots include in the object
}

\author{ Xuekui Zhang, Arnaud Droit <\email{arnaud.droit@crchuq.ualaval.ca}> and Raphael Gottardo <\email{rgottard@fhcrc.org}>}


\seealso{
\code{\link{summary}}
}

\keyword{print}
