\name{summary}
\Rdversion{1.1}
\docType{methods}
\alias{summary-methods}
\alias{summary}
\alias{summary,pics-method}
\alias{summary,picsList-method}
\alias{summary,segReads-method}
\alias{summary,segReadsList-method}
\title{summary PICS } 

\description{
This methods summary the objects of PICS.}

\usage{
\S4method{summary}{pics}(object)
\S4method{summary}{picsList}(object)
\S4method{summary}{segReads}(object)
\S4method{summary}{segReadsList}(object)
}


\arguments{
\item{object}{Object returned from \code{\link{pics}} .}
}


\author{
Xuekui Zhang <\email{xzhang@stat.ubc.ca}
Arnaud Droit, \email{arnaud.droit@ircm.qc.ca}
Raphael Gottardo <\email{raphael.gottardo@ircm.qc.ca}
  
}


\seealso{
\code{\link{show}}
}

\keyword{print}
