\name{unique}
\Rdversion{1.1}
\docType{methods}
\alias{unique-methods}
\alias{unique}
\alias{unique,GenomeData-method}

\title{GenomeData Unique Reads} 

\description{
This methods select the unique elements in a GenomeData object
}

\arguments{
\item{object}{A `GenomeData' object.}
}

\author{
Xuekui Zhang <\email{xzhang@stat.ubc.ca}
Arnaud Droit, \email{arnaud.droit@ircm.qc.ca}
Raphael Gottardo <\email{raphael.gottardo@ircm.qc.ca}
  
}

\seealso{
\code{\link{segmentReads}}
}

\keyword{print}