% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picsClass.R
\docType{class}
\name{pics-class}
\alias{pics-class}
\alias{show,pics-method}
\alias{minRange,pics-method}
\alias{maxRange,pics-method}
\alias{score,pics-method}
\alias{scoreReverse,pics-method}
\alias{scoreForward,pics-method}
\alias{chromosome,pics-method}
\alias{se,pics-method}
\alias{seF,pics-method}
\alias{seR,pics-method}
\alias{sigmaSqF,pics-method}
\alias{sigmaSqR,pics-method}
\alias{delta,pics-method}
\alias{mu,pics-method}
\alias{w,pics-method}
\alias{K,pics-method}
\alias{code,pics-method}
\alias{summary,pics-method}
\title{PICS class}
\usage{
\S4method{show}{pics}(object)

\S4method{minRange}{pics}(x)

\S4method{maxRange}{pics}(x)

\S4method{score}{pics}(x)

\S4method{scoreReverse}{pics}(x)

\S4method{scoreForward}{pics}(x)

\S4method{chromosome}{pics}(x)

\S4method{se}{pics}(x)

\S4method{seF}{pics}(x)

\S4method{seR}{pics}(x)

\S4method{sigmaSqF}{pics}(x)

\S4method{sigmaSqR}{pics}(x)

\S4method{delta}{pics}(x)

\S4method{mu}{pics}(x)

\S4method{w}{pics}(x)

\S4method{K}{pics}(x)

\S4method{code}{pics}(x)

\S4method{summary}{pics}(object)
}
\arguments{
\item{object, x}{A \code{pics} object.}
}
\description{
This object is used to gather all parameters from fitting PICS to a single candidate region.
}
\section{Functions}{
\itemize{
\item \code{show,pics-method}: show method

\item \code{minRange,pics-method}: Get start of range

\item \code{maxRange,pics-method}: Get end of range

\item \code{score,pics-method}: Score accessor.

\item \code{scoreReverse,pics-method}: Reverse score accessor.

\item \code{scoreForward,pics-method}: Forward score accessor.

\item \code{chromosome,pics-method}: Chromosome accessor

\item \code{se,pics-method}: se accessor

\item \code{seF,pics-method}: Forward se accessor

\item \code{seR,pics-method}: Reverse se accessor

\item \code{sigmaSqF,pics-method}: sigmaSqF accessor

\item \code{sigmaSqR,pics-method}: sigmaSqR accessor

\item \code{delta,pics-method}: delta accessor

\item \code{mu,pics-method}: mu accessor

\item \code{w,pics-method}: w accessor

\item \code{K,pics-method}: K accessor

\item \code{code,pics-method}: Error code accessor

\item \code{summary,pics-method}: Summary of the object
}}

\section{Slots}{

\describe{
\item{\code{estimates}}{A list containing all parameters estimates as well as standard errors.}

\item{\code{Nmerged}}{The number of binding events that were merged; binding events that overlap are merged.}

\item{\code{converge}}{A \code{logical} value indicating whether the EM algorithm has converged.}

\item{\code{chr}}{The candidate region's chromosome.}

\item{\code{score}}{Score of the binding event}

\item{\code{scoreF}}{Forward score of the binding event.}

\item{\code{scoreR}}{Reverse score of the binding event.}

\item{\code{range}}{Genomic ranges.}
}}

