% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PICS.R
\name{PICS}
\alias{PICS}
\title{Estimation of binding site positions}
\usage{
PICS(
  segReadsList,
  dataType = NULL,
  paraEM = NULL,
  paraPrior = NULL,
  nCores = 1
)
}
\arguments{
\item{segReadsList}{This object contains segmentation of Genome}

\item{dataType}{A \code{character}. The type of data you are processing: 
'TF' for transcription factor.}

\item{paraEM}{A \code{list} of parameters for the EM algorithm as returned by 
the \code{setParaEM} function. The default parameters should be good enough 
for most usages.}

\item{paraPrior}{A \code{list} of parameters for the priors as returned by
\code{setParaPrior}.}

\item{nCores}{An \code{integer}. The number of cores that should be used in
parallel by the function.}
}
\value{
An object of class \code{picsList} containing the estimated binding
site positions.
}
\description{
This object contains Estimation of binding site positions and has the
following slots: segReadsList, dataType.
}
