% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picsClass.R
\docType{class}
\name{picsError-class}
\alias{picsError-class}
\alias{show,picsError-method}
\alias{minRange,picsError-method}
\alias{maxRange,picsError-method}
\alias{score,picsError-method}
\alias{scoreReverse,picsError-method}
\alias{scoreForward,picsError-method}
\alias{chromosome,picsError-method}
\alias{se,picsError-method}
\alias{seF,picsError-method}
\alias{seR,picsError-method}
\alias{sigmaSqF,picsError-method}
\alias{sigmaSqR,picsError-method}
\alias{delta,picsError-method}
\alias{mu,picsError-method}
\alias{w,picsError-method}
\alias{K,picsError-method}
\alias{code,picsError-method}
\title{picsError class}
\usage{
\S4method{show}{picsError}(object)

\S4method{minRange}{picsError}(x)

\S4method{maxRange}{picsError}(x)

\S4method{score}{picsError}(x)

\S4method{scoreReverse}{picsError}(x)

\S4method{scoreForward}{picsError}(x)

\S4method{chromosome}{picsError}(x)

\S4method{se}{picsError}(x)

\S4method{seF}{picsError}(x)

\S4method{seR}{picsError}(x)

\S4method{sigmaSqF}{picsError}(x)

\S4method{sigmaSqR}{picsError}(x)

\S4method{delta}{picsError}(x)

\S4method{mu}{picsError}(x)

\S4method{w}{picsError}(x)

\S4method{K}{picsError}(x)

\S4method{code}{picsError}(x)
}
\arguments{
\item{object, x}{A \code{picsError} object.}
}
\description{
This class is used in cases when the algorithm does not converge.
}
\section{Functions}{
\itemize{
\item \code{show,picsError-method}: show method

\item \code{minRange,picsError-method}: Get start of range

\item \code{maxRange,picsError-method}: Get end of range

\item \code{score,picsError-method}: Score accessor.

\item \code{scoreReverse,picsError-method}: Reverse score accessor.

\item \code{scoreForward,picsError-method}: Forward score accessor.

\item \code{chromosome,picsError-method}: Chromosome accessor

\item \code{se,picsError-method}: se accessor

\item \code{seF,picsError-method}: Forward se accessor

\item \code{seR,picsError-method}: Reverse se accessor

\item \code{sigmaSqF,picsError-method}: sigmaSqF accessor

\item \code{sigmaSqR,picsError-method}: sigmaSqR accessor

\item \code{delta,picsError-method}: delta accessor

\item \code{mu,picsError-method}: mu accessor

\item \code{w,picsError-method}: w accessor

\item \code{K,picsError-method}: K accessor

\item \code{code,picsError-method}: Error code accessor
}}

\section{Slots}{

\describe{
\item{\code{errorCode}}{The error code for debugging.}
}}

