% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picsFDR.R
\name{picsFDR}
\alias{picsFDR}
\title{Estimate the FDR}
\usage{
picsFDR(
  picsIP,
  picsCont,
  filter = list(delta = c(0, Inf), se = c(0, Inf), sigmaSqF = c(0, Inf), sigmaSqR = c(0,
    Inf))
)
}
\arguments{
\item{picsIP}{An object of class \code{pics} or \code{picsList} containing
the informations for the IP reads.}

\item{picsCont}{An object of class \code{pics} or \code{picsList} containing
the informations for the control reads}

\item{filter}{A \code{list} of ranges for filtering regions based on PICS
parameters. By default filter is set to \code{NULL} and all regions are used.
\itemize{
  \item{delta:} Length of the binding sites
  \item{se:} Standard error
  \item{sigmaSqF:} Forward peak variance
  \item{sigmaSqR:} Reverse peak variance
}}
}
\value{
A \code{data.frame} with the following columns: FDR, score, N
}
\description{
Estimate the false detection rate for an object of class \code{pics} or 
\code{picsList}.
}
\seealso{
picsList pics
}
