% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picsClass.R
\docType{class}
\name{picsList-class}
\alias{picsList-class}
\alias{show,picsList-method}
\alias{minRange,picsList-method}
\alias{maxRange,picsList-method}
\alias{score,picsList-method}
\alias{scoreReverse,picsList-method}
\alias{scoreForward,picsList-method}
\alias{chromosome,picsList-method}
\alias{se,picsList-method}
\alias{seF,picsList-method}
\alias{seR,picsList-method}
\alias{sigmaSqF,picsList-method}
\alias{sigmaSqR,picsList-method}
\alias{delta,picsList-method}
\alias{mu,picsList-method}
\alias{w,picsList-method}
\alias{K,picsList-method}
\alias{code,picsList-method}
\alias{length,picsList-method}
\alias{summary,picsList-method}
\alias{[,picsList,ANY,ANY,ANY-method}
\alias{[[,picsList,ANY,ANY-method}
\title{List of PICS objects}
\usage{
\S4method{show}{picsList}(object)

\S4method{minRange}{picsList}(x)

\S4method{maxRange}{picsList}(x)

\S4method{score}{picsList}(x)

\S4method{scoreReverse}{picsList}(x)

\S4method{scoreForward}{picsList}(x)

\S4method{chromosome}{picsList}(x)

\S4method{se}{picsList}(x)

\S4method{seF}{picsList}(x)

\S4method{seR}{picsList}(x)

\S4method{sigmaSqF}{picsList}(x)

\S4method{sigmaSqR}{picsList}(x)

\S4method{delta}{picsList}(x)

\S4method{mu}{picsList}(x)

\S4method{w}{picsList}(x)

\S4method{K}{picsList}(x)

\S4method{code}{picsList}(x)

\S4method{length}{picsList}(x)

\S4method{summary}{picsList}(object)

\S4method{[}{picsList,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[[}{picsList,ANY,ANY}(x, i, j, ..., exact = TRUE)
}
\arguments{
\item{object, x}{A \code{pics} object.}

\item{i, j, ..., drop, exact}{Arguments passed to subset functions}
}
\description{
List of PICS objects
}
\section{Functions}{
\itemize{
\item \code{show,picsList-method}: show method

\item \code{minRange,picsList-method}: Get start of range

\item \code{maxRange,picsList-method}: Get end of range

\item \code{score,picsList-method}: Score accessor.

\item \code{scoreReverse,picsList-method}: Reverse score accessor.

\item \code{scoreForward,picsList-method}: Forward score accessor.

\item \code{chromosome,picsList-method}: Chromosome accessor

\item \code{se,picsList-method}: se accessor

\item \code{seF,picsList-method}: Forward se accessor

\item \code{seR,picsList-method}: Reverse se accessor

\item \code{sigmaSqF,picsList-method}: sigmaSqF accessor

\item \code{sigmaSqR,picsList-method}: sigmaSqR accessor

\item \code{delta,picsList-method}: delta accessor

\item \code{mu,picsList-method}: mu accessor

\item \code{w,picsList-method}: w accessor

\item \code{K,picsList-method}: K accessor

\item \code{code,picsList-method}: Error code accessor

\item \code{length,picsList-method}: Return the length of the object

\item \code{summary,picsList-method}: Summary of the object

\item \code{[,picsList,ANY,ANY,ANY-method}: Subset list

\item \code{[[,picsList,ANY,ANY-method}: Subset element
}}

