% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethods.R
\name{plot-FDR}
\alias{plot-FDR}
\alias{plot,picsList,picsList-method}
\title{Plot PICS FDR}
\usage{
\S4method{plot}{picsList,picsList}(x, y, filter = NULL, h = 0.1, ...)
}
\arguments{
\item{x}{A \code{picsList} object as returned by the function \code{PICS} run on the treatment data.}

\item{y}{A \code{picsList} object as returned by the function \code{PICS} run on the control data.}

\item{filter}{A list of ranges for filtering regions based on \code{PICS} parameters. By default filter is set to 'NULL' and all regions are used.}

\item{h}{A value between 0 and 1, representing the desired FDR. This simply draws a horizontal line at the given value.}

\item{...}{Further graphical parameters passed to the generic function \code{plot}.}
}
\description{
This method plots a curve showing the FDR as a function of the PICS scores.
}
\seealso{
\code{PICS}
}
