% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segReadsClass.R
\docType{class}
\name{segReads-class}
\alias{segReads-class}
\alias{segReads}
\alias{show,segReads-method}
\alias{map,segReads-method}
\alias{summary,segReads-method}
\title{Classe and methods for candidate regions}
\usage{
segReads(yF, yR, cF, cR, map, chr)

\S4method{show}{segReads}(object)

\S4method{map}{segReads}(object)

\S4method{summary}{segReads}(object)
}
\arguments{
\item{yF}{A \code{numeric} vector. Forward reads.}

\item{yR}{A \code{numeric} vector. Reverse reads.}

\item{cF}{A \code{numeric} vector. Forward reads for the controls.}

\item{cR}{A \code{numeric} vector. Reverse reads for the controls.}

\item{map}{A \code{matrix}. The mappability profile.}

\item{chr}{A \code{character}. Chromosome name.}

\item{object}{A \code{segReads} object.}
}
\description{
A segReads object represents a single candoidate region, including all its
informative reads and mappability profile.
}
\section{Functions}{
\itemize{
\item \code{segReads}: \code{segReads} constructor

\item \code{show,segReads-method}: show method

\item \code{map,segReads-method}: map method

\item \code{summary,segReads-method}: summary method
}}

\note{
segReads objects are not meant to be built via the
constructors. The constructors are used in \code{segmentPICS}.
}
