% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segReadsClass.R
\docType{class}
\name{segReadsList-class}
\alias{segReadsList-class}
\alias{segReadsList}
\alias{show,segReadsList-method}
\alias{map,segReadsList-method}
\alias{summary,segReadsList-method}
\alias{[,segReadsList,ANY,ANY,ANY-method}
\alias{[[,segReadsList,ANY,ANY-method}
\title{Class and methods for list of candidate regions}
\usage{
segReadsList(List, paraSW, N, Nc)

\S4method{show}{segReadsList}(object)

\S4method{map}{segReadsList}(object)

\S4method{summary}{segReadsList}(object)

\S4method{[}{segReadsList,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[[}{segReadsList,ANY,ANY}(x, i, j, exact = TRUE)
}
\arguments{
\item{List}{A \code{list} of \code{segReads} objects.}

\item{paraSW}{A \code{list} of parameters for the genomic regions.}

\item{N}{A \code{numeric}. The number of reads in the data.}

\item{Nc}{A \code{numeric}. The number of reads in the control.}

\item{object}{A \code{segReadsList} object.}

\item{x, i, j, ..., drop, exact}{Arguments passed to subset methods}
}
\description{
Class and methods for list of candidate regions
}
\section{Functions}{
\itemize{
\item \code{segReadsList}: \code{segReadsList} constructor

\item \code{show,segReadsList-method}: show method

\item \code{map,segReadsList-method}: map method

\item \code{summary,segReadsList-method}: summary method

\item \code{[,segReadsList,ANY,ANY,ANY-method}: Subset methods

\item \code{[[,segReadsList,ANY,ANY-method}: Subset methods
}}

\note{
segReadsList objects are not meant to be built via the
constructors. The constructors are used in \code{segmentPICS}.
}
