% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segReadsPEClass.R
\docType{class}
\name{segReadsListPE-class}
\alias{segReadsListPE-class}
\alias{segReadsListPE}
\alias{[,segReadsListPE,ANY,ANY,ANY-method}
\alias{[[,segReadsListPE,ANY,ANY-method}
\title{Class and methods for list of candidate regions from paired-end data}
\usage{
segReadsListPE(List, paraSW, N, NFm, NRm, Nc, NcFm, NcRm)

\S4method{[}{segReadsListPE,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[[}{segReadsListPE,ANY,ANY}(x, i, j, ..., exact = TRUE)
}
\arguments{
\item{x, i, j, ..., drop, exact}{Arguments passed to subset methods}
}
\description{
Class and methods for list of candidate regions from paired-end data
}
\section{Functions}{
\itemize{
\item \code{segReadsListPE}: \code{segReadsListPE} constructor.

\item \code{[,segReadsListPE,ANY,ANY,ANY-method}: subset method

\item \code{[[,segReadsListPE,ANY,ANY-method}: subset method
}}

