\name{segReadsListPE}
\alias{segReadsListPE}
\alias{segReadsListPE-class}
\alias{[,segReadsListPE-method}
\alias{[,segReadsListPE,ANY,ANY-method}
\alias{[[,segReadsListPE,ANY,ANY-method}

\title{List of segReadsPE objects}

\description{
A list of segReadsPE. The class also store information related to the segmentation process, keeping a trace of the parameters used and the proportion of forward and reverse reads for the input and the control.
}

\section{Methods}{
  \describe{
 \item{[}{\code{signature(x = ``pics'')}: subset gadem object.}
 \item{[[}{\code{signature(x = ``pics'')}: subset gadem object.}
	}
}


\section{Extends}{
Class \code{\linkS4class{segReadsList}}, directly.
}

\author{Xuekui Zhang}

\seealso{
  \code{\linkS4class{segReadsPE}}
  \code{\linkS4class{segReadsList}}
}

\keyword{classes}
