% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segReadsPEClass.R
\docType{class}
\name{segReadsPE-class}
\alias{segReadsPE-class}
\alias{segReadsPE}
\title{Classe and methods for candidate regions from paired-end data}
\usage{
segReadsPE(yF, yR, yFm, yRm, cF, cR, cFm, cRm, map, chr)
}
\description{
A segReadsPE object represents a single candoidate region, including all its
informative reads and mappability profile.
}
\section{Functions}{
\itemize{
\item \code{segReadsPE}: \code{segReadsPE} constructor.
}}

