\name{summary}
\Rdversion{1.1}
\docType{methods}
\alias{summary-methods}
\alias{summary}
\alias{summary,pics-method}
\alias{summary,picsList-method}
\alias{summary,segReads-method}
\alias{summary,segReadsList-method}
\title{summary PICS } 

\description{
This methods summarized `pics', `picsList', `seg' or `segList' objects.}

\usage{
\S4method{summary}{pics}(object)
\S4method{summary}{picsList}(object)
\S4method{summary}{segReads}(object)
\S4method{summary}{segReadsList}(object)
}


\arguments{
\item{object}{Object returned from \code{\link{pics}} .}
}


\author{ Xuekui Zhang, Arnaud Droit <\email{arnaud.droit@crchuq.ualaval.ca}> and Raphael Gottardo <\email{rgottard@fhcrc.org}>}



\seealso{
\code{\link{show}}
}

\keyword{print}
