% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segReadsGeneric.R
\name{summarySeg}
\alias{summarySeg}
\title{Summarize segmentList objects}
\usage{
summarySeg(seg)
}
\arguments{
\item{seg}{A \code{segmentList} object as returned by \code{segmentPICS}.}
}
\value{
A \code{data.frame}. With 
 \itemize{
   \item{chr:} Chromosome id
   \item{NF:} Number of forward reads
   \item{NR:} Number of reverse reads
   \item{L:} Length of segment
   \item{min:} Start location of segments
   \item{max:} End location of segments
 }
}
\description{
Summarize segmentList objects into a data.frame
}
